Akintor/requirements.txt
cloudscraper>=1.2.60
beautifulsoup4>=4.11.1
pyyaml>=6.0
